Working with Nested Dictionaries and Lists
I created this exercise after working with Google's Directory API users.aliases.list() method.

The server response of this method is a JSON object that can be interpreted as a dictionary, one of its keys is a list of multiples Dictionaries, and I am only interested in the value of one of those keys.

I have already been able to list all user alises from the response of the users.list() method but I wanted to work directly with the users.aliases.list() but it turned out to be more complicated. My original approach can be seen here: https://github.com/jnyada/Learning-Google-Workspace-API/blob/master/options.py.

The idea behind the code is, to check if the 'aliases' key is included or not, since the response includes multiple JSONs and not all of them have it. After it is found to read the list and print just the one specific alias or aliases of the user.

In this excersie I am looking to print the value of the key "I" which is "i"